requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# LangChain ecosystem
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.10

# Vector store
chromadb>=0.4.0

# Embeddings
google-generativeai>=0.3.0
sentence-transformers>=2.2.0

# Schema validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Tokenization
tiktoken>=0.5.0

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.0.0

# FastAPI for web service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Streamlit for UI
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0
